Move all evaluators and relative utility in one class
