--- hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/source/evaluator/AlwaysFalse.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.hudi.source.evaluator;
-
-import org.apache.hudi.source.stats.ColumnStats;
-
-import java.util.Map;
-
-/**
- * A special evaluator which is not possible to match any condition.
- */
-public class AlwaysFalse implements Evaluator {
-  private static final long serialVersionUID = 1L;
-
-  public static AlwaysFalse getInstance() {
-    return new AlwaysFalse();
-  }
-
-  @Override
-  public boolean eval(Map<String, ColumnStats> columnStatsMap) {
-    return false;
-  }
-}
